isu/requirements.txt
asgiref==3.7.1
certifi==2023.5.7
Django==4.2.1
django-elasticsearch-dsl==7.3
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==1.26.16
celery[redis]==5.2.7
flower==1.2.0
